video-bot/requirements.txt
torch==2.2.2
torchaudio==2.2.2
numpy==1.26.4
librosa==0.9.2
pydub==0.25.1
soundfile==0.12.1
pyannote.audio==3.2.1
einops==0.8.1
huggingface-hub==0.23.1
